Refactoring schematics folder and file structure
