+++ packages/schematics/src/rules/install-node-packages.rule.ts
@@ -1,0 +1,10 @@
+import { Rule, SchematicContext, Tree } from '@angular-devkit/schematics';
+import { NodePackageInstallTask } from '@angular-devkit/schematics/tasks';
+
+export function installDependencies(): Rule {
+  return (tree: Tree, context: SchematicContext) => {
+    context.addTask(new NodePackageInstallTask());
+    context.logger.info('✅️ Dependencies installed');
+    return tree;
+  };
+}
