+++ packages/schematics/src/schematics/ng-add/files/main.augury.ts
