+++ packages/schematics/src/schematics/ng-add/schema.json
