+++ packages/schematics/src/types/ng-add.schema.ts
@@ -1,4 +1,4 @@
-export interface Schema {
+export interface NgAddSchema {
   /** Name of the project to target. */
   project: string;
 }
