+++ packages/schematics/src/types/npm-registry-package.ts
@@ -1,0 +1,4 @@
+export interface NpmRegistryPackage {
+  name: string;
+  version: string;
+}
