+++ packages/schematics/src/types/npm-script.ts
@@ -1,0 +1,4 @@
+export interface NpmScript {
+  key: string;
+  command: string;
+}
